# store: Optimized common cases for time selecting smaller amount of series. Avoid looking up symbols.